First stable version of the billiard system.
Still have the problem that near corners particles find infinite collision time, but this is the only "problem". Everything else is completely stable.

Of course the above is not a problem, but a trade-off since you must not have particles that have very small collision times because the side decider can be wrong.